complete authorization with firebase
