+++ Config/Dev/salon-cloud-dev-81007-firebase-adminsdk-0lbta-29b3ebe135.json
@@ -1,0 +1,12 @@
+{
+  "type": "service_account",
+  "project_id": "salon-cloud-dev-81007",
+  "private_key_id": "29b3ebe135d43fa050331816706db7935cef1eb5",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDUc5o1qsKTCYfM\n1ioV0ALGApPtTNRGjFgFSqRzn2un4PlaTJKn1kLp2BuEHXuQ7cdXzk/j1eYUpHXy\nsi2MkbyYxBUEbtltntbYTfiZ4qyQwDgpeY2iBEmgCVIT+zTM9bjv05m6ryN4E7mU\nKl6R2eVt4NGYWST4UdvLbbwcGynuEtzT4qN4Yn2ngrtBbpfAYr+iq59IWRXetUH6\nE0O68AofRHMi25chcLfA6/dboreze6zPWbYZNywlCsqjX+ZyNwcRJLkd++E2RWK2\nVojD/yYdvk1vgIjCz6xdwrnjokpfS0VmIHxzphi3R6RNp67fepsHGS/GsKY4hOGI\nVB2cE3U7AgMBAAECggEAFMNbjeUrcP1cdUt0/3691qZH2MhKcr6il5nM6rPo7JZT\nKF0VdCu/jGUfg1DyU565fb5fFnBANUJBeL+rLfLMUM3RFAbVXx5H9wnfwwJhRSBU\nHNGRYwUyQN7IU8WYSnO1aFLw0krtl32uBGoDphN6lbuEIHSN5WrypTYLTV/mQCsQ\nFycnI/i44SJrOxJqruBPiA80pux7eK7MMcJ3X5ZmLbUdQA7G1TDOcP0bI5PRJd3L\nv6uid2569AjvT+5wO9EpaqcSAAjwLzH7B3C35yfAYRSXT+sSw4sxW++JOgGAPMzo\nEod9BGhPLdCtftLvHb28TDxpo8cBXYhMaX3qPz1sSQKBgQD4Ww1u2c+/oyt57SFW\nw5Es1LiSvtFjLRKuJ5zOURlojNz0phwenfp4YtywUaNEXXnbvTlOC3MB7M5xh2dl\nQTTBBTj68Qk/LnZGHJMkZT7C7kaczVazGtY2V2NfLisM4Nm4TKpBvsHeyI1tCHX7\nQl10tgN4IkaB0xqwrL4hqLCZlwKBgQDa/aOsKTr6Dkf491k23R+n+SBOHu2QCMvU\ngmD9dwmFT0j1qtSy1EsOVXwHBZAmAlCSD7aQTLgecsEmg56l89dCBGMkuzL5USUn\n0KaFWXMRBC67NG8IgwFMfgDSWow2zp17YDlgAGtgk83ePWIQ9Cv5/hTNk9d9tHJO\nheE3QhkN/QKBgQCXPauyg8X16AID4BpbYgE4kdpM0xNkGAC2Ob1ODykv+AhenAGn\ny4DKazoFzdFlQp+A53zMQECrPXCCxnh6uvjLSsjmWH7PERZJoAEjOty7jKtHtexE\naeVldlY3kj/5ObycdWOBYg9iAjPN4Eco2cNb3bPDJuxVBx/L04Z9rOLOxQKBgCjL\n0Z7z+2G0IWJXKg6XeRtAxOzoxBbuHJLVSx5M+8AS+iOSFJfQBr9/22AvgCAklD/x\nn3AYasB5XlQ/duGxoGtTB+lC2ocpv8NAEpWLCW2lGtOTPBAY37gRTTPi3HYlXZ8r\nWOTyjw5p2pNJWiiQZFTpKZT2stFfOWA5frBdpu89AoGBAIIqU9a8tZjsCj4G429q\n9KKH74bp3lZPxhgRjH4upkk+KhSIEqWdzWLfEthzXWahfDUMTNyEl4yaio4JIinQ\nEc9nvor4MtmN0TIqIrPO+iyuSIRUSAcxvtuSiQMYHB4J/ZaA5MJWTH86sy6Kwx91\nep1XevV1uHJR21pTBZ/cdhvK\n-----END PRIVATE KEY-----\n",
+  "client_email": "firebase-adminsdk-0lbta@salon-cloud-dev-81007.iam.gserviceaccount.com",
+  "client_id": "115576625760580026267",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://accounts.google.com/o/oauth2/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-0lbta%40salon-cloud-dev-81007.iam.gserviceaccount.com"
+}
